.claude/worktrees/stupefied-burnell/Tax-Advisor/requirements.txt
claude-agent-sdk==0.1.36
python-dotenv==1.2.1
